Fix for speed test decrease caused by loop.cycle
Noticed that the change in 2319405952cafa253e7dd688f1a3943fec7e033d that added the `loop.cycle` option dramatically reduced render speed. For some reason, defining a function on the `_context` reduces the speed. I'm not sure why--would be good to also investigate why that happens, as this is more or less doing the same thing, just adding the function to a different object.

Conflicts:

	lib/tags.js