Create SimpleTags -  a way to easily create tags that generate html.
While custom tags are very powerful, generating a function that generates
html isn't particularly easy or clean for the fairly common case of just
wanting a tag to generate some html.

This allows you to specify simpleTags - functions that return html - and
generates tags from them.

If it's accepted I'm happy to write docs for this.